pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dle-saas"
version = "0.1.0"
description = "Initial quality baseline and scaffolding for DLE-SaaS."
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2,<5.3",
    "djangorestframework>=3.16,<3.17",
    "drf-spectacular>=0.29,<0.30",
    "gunicorn>=23.0,<24.0",
    "psycopg[binary]>=3.2,<3.3",
]

[tool.setuptools]
packages = []

[project.optional-dependencies]
dev = [
    "bandit>=1.7",
    "django-stubs>=5.0",
    "factory-boy>=3.3",
    "mypy>=1.10",
    "pip-audit>=2.8",
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-django>=4.8",
    "ruff>=0.11",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "_research",
]

[tool.ruff.lint]
select = [
    "B",
    "E",
    "F",
    "I",
    "RUF",
    "SIM",
    "UP",
]

[tool.ruff.lint.per-file-ignores]
"backend/tests/*.py" = ["S101"]
"backend/apps/*/migrations/*.py" = ["E501", "RUF012"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
files = ["backend"]
mypy_path = ["backend"]
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
pretty = true
show_error_codes = true
exclude = ["^_research/"]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["backend.tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["rest_framework", "rest_framework.*", "drf_spectacular", "drf_spectacular.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
python_files = ["test_*.py"]
testpaths = ["backend"]
pythonpath = ["backend"]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "--maxfail=1",
]

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[tool.coverage.run]
branch = true
source = ["backend"]

[tool.coverage.report]
show_missing = true
skip_covered = true
